+++ packages/core/src/components/icon-button/styles.ts
@@ -4,7 +4,7 @@ export const root = cva(
   [
     "inline-flex aspect-square h-10 items-center justify-center rounded-full text-icon-medium",
     "outline-none state-layer ripple reset-svg",
-    "ui-hover:hover-state-layer ui-focus:focus-state-layer ui-focus-visible:hover-state-layer ui-active:press-state-layer",
+    "focus:focus-state-layer focus-visible:hover-state-layer ui-not-disabled:hover:hover-state-layer ui-not-disabled:active:press-state-layer",
   ],
   {
     variants: {
@@ -14,7 +14,7 @@ export const root = cva(
           "bg-secondary-container text-on-secondary-container ui-disabled:bg-on-surface/12 ui-disabled:text-on-surface/38",
         ],
         outlined: [
-          "border border-outline bg-surface text-primary ui-disabled:border-on-surface/12 ui-disabled:text-on-surface/38 ui-focus:border-primary",
+          "border border-outline bg-surface text-primary focus:border-primary ui-disabled:border-on-surface/12 ui-disabled:text-on-surface/38",
         ],
         standard: ["bg-transparent text-on-surface-variant ui-disabled:text-on-surface/38"],
       },