+++ packages/core/src/components/navigation-drawer-item/styles.ts
@@ -4,7 +4,7 @@ import { cva } from "class-variance-authority";
 export const root = cva(
   [
     "inline-flex h-14 w-full select-none items-center gap-x-3 rounded-full pl-3 pr-6 outline-none state-layer ripple",
-    "ui-hover:hover-state-layer ui-focus:focus-state-layer ui-focus-visible:hover-state-layer ui-active:press-state-layer",
+    "focus:focus-state-layer focus-visible:hover-state-layer ui-not-disabled:hover:hover-state-layer ui-not-disabled:active:press-state-layer",
     "ui-in-route:bg-secondary-container ui-in-route:text-on-secondary-container ui-not-in-route:text-on-surface-variant",
   ],
   {